+++ R/rice.R
@@ -13,13 +13,15 @@
 
 # error multipliers, rounding. Could add procedures for different labs, e.g. QUB_bg, etc. This would be useful for reasons of transparency and community standards. Add data from historical UBA standards/backgrounds?
 
+
+
 #' @name howmuchC14
 #' @title Amount of C14 particles in a sample
 #' @description Calculate the expected amount of remaining C14 atoms in a sample, given its weight and age.
 #' @details The number of carbon atoms in the sample is estimated. Given the known C14/C ratio at F=1, and given the sample's age, we can estimate the number of remaining C14 atoms. Given a 12C current at the detector end of an AMS, we can then also calculate how many 14C ions would be counted per second and minute. 
-#' Note that backgrounds are not modelled (but could be investigated by e.g. typing \code{howmuchC14(45e3)} which gives as c. 1 background count per second).  
+#' Note that backgrounds are not modelled (but could be investigated by e.g. typing \code{howmuchC14(45e3)} which gives as c. 1 background count per second). The calculated C14 count rate assumes no fractionation.
 #' @return The estimated number of C14 atoms.
-#' @param age The age of the sample (in cal BP per default, or in C14 BP is use.cc=FALSE).
+#' @param age The age of the sample (in cal BP per default, or in C14 BP if use.cc=FALSE).
 #' @param wght The weight of the sample (in mg). Defaults to 1 mg.
 #' @param use.cc Whether or not to use the calibration curve. If set to \code{use.cc=FALSE}, then we assume that the age is the radiocarbon age (this enables ages beyond the reach of the calibration curves to be used).
 #' @param Av Avogadro's number, used to calculate the number of carbon atoms in the sample.
@@ -28,6 +30,7 @@
 #' @param format The format of the printed numbers. Defaults to either scientific (for large numbers) or as fixed-point, depending on the size of the number.
 #' @param cc calibration curve for C14 (see \code{caldist()}).
 #' @param postbomb Whether or not to use a postbomb curve (see \code{caldist()}).
+#' @param glue Glue postbomb and prebomb curves together. Defaults to 0 (none), can be 1 (IntCal20 + NH1), 2 (IntCal20 + NH2), 3 (IntCal20 + NH3), 4 (SHCal20 + SH1-2) or 5 (SHCal20 + SH3). Note that this will override the value of cc. Can also be used to glue other pre-bomb and post-bomb curves, e.g., \code{glue="NH1_monthly"}.
 #' @param cc.dir Directory of the calibration curves. Defaults to where the package's files are stored (system.file), but can be set to, e.g., \code{cc.dir="curves"}.
 #' @param thiscurve As an alternative to providing cc and/or postbomb, the data of a specific curve can be provided (3 columns: cal BP, C14 age, error).
 #' @param talk Whether or not to provide feedback (defaults to TRUE).
@@ -38,16 +41,18 @@
 #'   howmuchC14(55e3) # at dating limit
 #'   howmuchC14(145e3) # way beyond the dating limit, 1 C14 atom per mg remains
 #' @export
-howmuchC14 <- function(age, wght=1, use.cc=TRUE, Av=6.02214076e23, C14.1950=1.176e-12, current=25e-6, format="g", cc=1, postbomb=FALSE, cc.dir=NULL, thiscurve=NULL, talk=TRUE, decimals=3) {
+howmuchC14 <- function(age, wght=1, use.cc=TRUE, Av=6.02214076e23, C14.1950=1.176e-12, current=25e-6, format="g", cc=1, postbomb=FALSE, glue=0, cc.dir=NULL, thiscurve=NULL, talk=TRUE, decimals=3) {
+  if(length(age) > 1)
+    stop("can only handle single value for age")
 
   if(use.cc) {
-    F <- calBPtoF14C(age, cc=cc, postbomb=postbomb, cc.dir=cc.dir, thiscurve=thiscurve)[,1]
+    F <- calBPtoF14C(age, cc=cc, postbomb=postbomb, glue=glue, cc.dir=cc.dir, thiscurve=thiscurve)[,1]
     if(is.na(F)) {
       message("Cannot use calibration curve for this age, assuming C14 age")
       F <- C14toF14C(age)
     }
   } else
-       F <- C14toF14C(age) # then t is on the C14 scale
+       F <- C14toF14C(age) # then 'age' is on the C14 scale
 
   F <- as.numeric(F)
   atoms <- (wght/1e3)*Av/12 # number of C atoms in a mg
@@ -74,7 +79,75 @@ howmuchC14 <- function(age, wght=1, use.cc=TRUE, Av=6.02214076e23, C14.1950=1.17
       60*persecond, " per minute)" ))
   }
 
-  invisible(C14)
+  invisible(list(C14=C14, C14persecond=i14))
+}
+
+
+
+#' @name radio
+#' @title Listen to radiocarbon being measured
+#' @description A sonification of the amount of C14 being detected in an AMS. 
+#' @details The sonification can be heard as 'Geiger counter-like' clicks (individual C14 atoms ending up in the detector) and/or as a possibly slightly meandering sine wave (e.g., for a sample of age 5000 14C BP, we'd count c. 105.6 atoms per second in the AMS, so this would become a tone of frequency 105.6 Hz). The calculations are based on the function \code{howmuchC14}.
+#' @return A sound (tone and or clicks)
+#' @param age The age of the sample (in cal BP per default, or in C14 BP if use.cc=FALSE).
+#' @param duration How long the sample will sound for in seconds. Defaults to 10 seconds.
+#' @param sr Sampling rate. This defaults to 44100 (per second), which is based on the CD standard.
+#' @param as.clicks Make the C-14 counts sound as clicks, based on random Poisson sampling.
+#' @param as.tone Make the C-14 frequency (counts per second) sound as a wave (e.g., 105 cps becomes a 105 Hz sine wave). This can either be a constant wave or be meandering (see `noise` and `meander`).
+#' @param click_length Length of the clicks. Defaults to 80, but can be altered to make the clicks sound different
+#' @param tone.volume Volume of the tone/wave relative to that of the clicks.
+#' @param noise Deviation from the mean of the tone. Defaults to 0.02.
+#' @param meander Drift of the tone along the mean. Defaults to 0.005. 
+#' @param ... Optional constants to be entered into the function `howmuchC14`
+#' @author Maarten Blaauw
+#' @examples
+#'   radio(0)
+#'   radio(45000)
+#' @export
+radio <- function(age, duration=10, sr=44100, as.clicks=TRUE, as.tone=TRUE, click_length=80, tone.volume=0.5, noise=0.02, meander=0.005, ...) {
+
+  hasaudio <- requireNamespace("audio", quietly=TRUE)
+  if(!hasaudio)
+    stop("Please install the audio package:\ninstall.packages(\"audio\")")	
+	
+  rate <- howmuchC14(age, ...)$C14persecond # C14 atoms counted per second
+  n <- sr * duration # number of steps
+  time <- 1:(n)
+ 
+  p <- rate / sr # probability per sample
+  events <- runif(n) < p # random Bernoulli trials
+
+  # make random clicks based on the rate
+  click <- rnorm(click_length) * exp(-seq(0, 6, length.out = click_length))
+  click <- click / max(abs(click))
+  clicks <- numeric(n)
+  for(i in which(events)) {
+    end <- min(i + click_length - 1, n)
+    clicks[i:end] <- clicks[i:end] + click[1:(end - i + 1)]
+  }
+  if(max(abs(clicks)) > 0)
+    clicks <- clicks / max(abs(clicks)) # normalise
+  
+  # make a sine wave based on the rate, perhaps wobble it a bit
+  tone <- sin(2 * pi * rate * time / sr)
+  if(noise > 0) {
+	noise <- rnorm(n, 0, noise * sqrt(rate))
+	noise[is.na(noise)] <- 0
+	wobble <- numeric(n)
+	wobble[1] <- rate 
+    for(i in 2:n)
+      wobble[i] <- max(0, wobble[i-1] + meander * (rate - wobble[i-1]) + noise[i])
+	tone <- sin(cumsum(2 * pi * wobble / sr))
+  }
+  
+  if(as.clicks && as.tone)
+    sound <- rbind(tone.volume*tone, clicks) else
+      if(as.clicks)
+        sound <- clicks else 
+          sound <- tone
+	   
+  audio::play(audio::audioSample(sound, sr))
+  invisible(sound)
 }